updated: unit test refactored
