+++ test/helpers/index.js
@@ -34,7 +34,6 @@ export function getNextSibling(n) {
   return x
 }
 
-
 export function getRiotStyles(riot) {
   // util.injectStyle must add <style> in head, not in body -- corrected
   var stag = riot.util.styleNode || document.querySelector('style')
@@ -61,87 +60,6 @@ export function $(selector, ctx) {
   return (ctx || document).querySelector(selector)
 }
 
-export function injectScript(path) {
-  var script = document.createElement('script')
-  script.src = path
-  script.type = 'riot/tag'
-  document.head.appendChild(script)
-}
-
-export function appendTag(name, attrs, innerHTML) {
-  var tag = document.createElement(name)
-  document.body.appendChild(tag)
-  if (attrs)
-    Object.keys(attrs).forEach(function(key) {
-      tag.setAttribute(key, attrs[key])
-    })
-  if (innerHTML)
-    tag.innerHTML = innerHTML
-  return tag
-}
-
-
-export function loadTagsAndScripts(arr) {
-
-  var tagsNames = []
-
-  arr.forEach(function(item) {
-    var path,
-      name
-
-    // if it's a string we assume it's just the path
-    if (typeof item === 'string') {
-      path = item
-    } else {
-      // get the item properties
-      path = item.path
-      name = item.name
-    }
-
-    // if name is undefined
-    // we assume the file name without extension is tag name
-    if (typeof name === 'undefined')
-      name = path.split('/').pop().replace(/(\..*|~)/g, '')
-
-    if (path)
-      injectScript(path)
-    if (name) {
-      appendTag(name, item.attrs, item.html)
-      tagsNames.push(name)
-    }
-  })
-
-  return tagsNames.join(',')
-}
-
-export function defineTag(tagDef) {
-  var name = tagDef.match(/^<([\w\-]+)/)[1]
-
-  // compile expects the closing tag not to have any leading whitespace
-  tagDef = tagDef.replace(/(\s*)(<\/[\w\-]+>)\s*$/, '\n$2')
-
-  riot.compile(tagDef)
-
-  // store the name so it can be un-registered
-  defineTag.names = defineTag.names || []
-  defineTag.names.push(name)
-}
-
-export function makeTag(htmlOrName, html) {
-  var name = html ? htmlOrName : 'test-tag'
-  if (!html) html = htmlOrName
-
-  html = '<' + name + '>\n' + html.trim() + '\n</' + name + '>'
-  defineTag(html)
-  injectHTML('<' + name + '/>')
-  var tags = riot.mount(name)
-
-  // store the tag, so it can be removed
-  makeTag.tags = makeTag.tags || []
-  makeTag.tags.push(tags[0])
-
-  return tags[0]
-}
 
 export function injectHTML(html) {
   var div = document.createElement('div')
